Fix some silly bugs in the recently committed code.
Not counting the number of occurrences of '\002' and '\003' in run_query_html() was resulting in insufficient memory allocation which caused segfaults for some queries (eg. 'copy strings', 'add user system' which are part of my adhoc set of queries I use for testing).

Prevent an infinite loop in run_query_pager was being triggered for some queries (like the ones mentioned abobe).