Write code for handling the Xr macros embedded inside NAME section of mdoc pages.
While the parsing logic for NAME section works fine in most of the cases but the Xr macros require special attention. Similar code was added for handling Xr macros in the body of the page but because the NAME section's parsing logic is separate, the new code had to be rewritten. Namely the pmdoc_Nd function contains the logic for recursing throught the nodes of the tree which are concerned with the NAME section. We use a static local variable in the function to remember if an Xr macro was seen recently, if yes, then we know that the current and the next node are the values of the Xr macro (namely the name of the man page and section number), so we parse those nodes and reset Xr = 0. Otherwise we have not seen an Xr node previously so we do the usual recursion and parsing.

Fixes #63