Initial support for parsing escape sequences.
Ideally this should be done in makemandb.c when storing the data in the db. But
for some reasons, it is not working well with the rest of the code in makemandb.

For now keep this code in apropos-utils.c. That is, until I figure out how to
make this code work in makemandb.c, pre-process the snippet to remove any
escape sequences in it.

Output looks much saner and readable in the absence of ugly escape sequences.