Strip down escape sequences while indexing.
This undoes the changes done in apropos-utils.c during past couple of commits.
Move the escape sequence parsing code from apropos-utils.c to makemandb.c.
Fix a couple of buffer overflows as well (thanks to Joerg).
Check for a \- separately as mandoc_parse does not seem to be able to handle it.

Remove dependency from libmandoc from rest of the source. Only makemandb depends on it.