Avoid an unnecessary call to memcpy for copying one byte.
Also, as we know in advance that the section number of a man page will take
only two bytes, declare it as char section[2], rather than allocating memory
for it on the heap each time.

Rename the functions get_section(), get_machine() to set_section(),
set_machinei() respectively. This seems more consistent, since we are setting
the values of rec->section and rec->machine.