Build a dictionary of the terms existing in the corpus.
Since, the main FTS table is using the Porter tokenizer, it is not
possible to use its index to extract the terms from it. Therefore,
the easiest and most efficient (determined experimentally) way to
build such a dictionary is to build another FTS index in an in-memory
db and then take out the words existing in the corpus using the fts4aux
module. It duplicates some part of the code but does the trick without
slowing down the indexing very much.