Spell check related fixes:
Since we are checking for stopword at multiple places, make it a
function
Convert a few ints to size_ts
In edits1, if only the new word is not a stop word add it to the
candidates list.
Since, the size of the candidates list returned by edits1 is no longer
deterministic, pass an int pointer to it, and edits1 will set it to the
appropride value to indicate the size of the resulting list.

Simplify the query in known_word, the subquery was unnecessary.

Don't check for known_word for the value passed to in the spell
function. We are calling spell to check for the alternatives, not to get
back the same value.